feat(android): P15 dialogo de confirmacion antes->despues al guardar
Antes de disparar el PATCH del empleado se muestra un MaterialAlertDialog
con el resumen "antes -> despues" de TODOS los campos modificados. El
ADMIN edita para corregir errores y este punto de confirmacion evita
guardar cambios involuntarios. Sin cambios -> Snackbar "Sin cambios" y
no se llama al API. Cancelar -> formulario intacto.

FormEmpleadoViewModel gana una funcion pura construirResumenCambios(estado)
que compara el estado actual contra los originales precargados de E15 y
devuelve la lista de campos modificados. El dni se compara normalizado
(trim + uppercase). fechaAlta lleva nota especial "Afecta a informes
historicos" porque su cambio repercute en saldos y reportes (M-037).

Como subproducto del refactor a EstadoFormulario se detecto un drift de
tipo en EmpleadoResponse.jornadaSemanalHoras (Int en cliente, Double en
backend y DDL desde v8). Se promueve el campo a Double en cliente y se
ajustan los lectores P14, P15 y P26 con un helper formatearHoras local
que muestra "40" si es entero y "37.5" si tiene decimales. El string
mi_perfil_jornada_valor pasa de %d a %s para aceptar la salida del helper.
Sin este fix una jornada parcial (37.5h) habria reventado en runtime con
IllegalFormatConversionException o se habria truncado silenciosamente al
deserializar. Drift documentado como M-038 en MEJORAS_V2.md.

Manual de usuario (B12 Anexo 4) sincronizado: menciona el resumen previo
al guardado y el aviso especial sobre la fecha de alta.

Tests Android pendientes en bloque para cuando se termine la
funcionalidad del TFG (cobertura actual: 1 test real).

+++ staffflow-android/app/src/main/java/com/staffflow/android/data/remote/dto/Responses.kt
@@ -62,7 +62,7 @@ data class EmpleadoResponse(
     val numeroEmpleado: String,
     val fechaAlta: String,
     val categoria: CategoriaEmpleado,
-    val jornadaSemanalHoras: Int,
+    val jornadaSemanalHoras: Double,
     val jornadaDiariaMinutos: Int,
     val diasVacacionesAnuales: Int,
     val diasAsuntosPropiosAnuales: Int,
